Trukket ut spesifikke hjemler til egne filer og fikset tester
