+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/hjemler/EØSForordningen883Hjemler.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.ba.sak.kjerne.brev.hjemler
+
+import no.nav.familie.ba.sak.kjerne.brev.domene.SanityEØSBegrunnelse
+
+fun hentEØSForordningen883Hjemler(
+    sanityEøsBegrunnelser: List<SanityEØSBegrunnelse>,
+) = sanityEøsBegrunnelser.flatMap { it.hjemlerEØSForordningen883 }.distinct()
