+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/brev/BrevServiceTest.kt
@@ -2,11 +2,9 @@ package no.nav.familie.ba.sak.kjerne.brev
 
 import io.mockk.every
 import io.mockk.mockk
-import no.nav.familie.ba.sak.common.TIDENES_ENDE
 import no.nav.familie.ba.sak.common.lagAndelTilkjentYtelse
 import no.nav.familie.ba.sak.common.lagBehandling
 import no.nav.familie.ba.sak.config.featureToggle.UnleashNextMedContextService
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseRepository
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.EndretUtbetalingAndelRepository
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.VedtaksperiodeService
@@ -14,9 +12,6 @@ import no.nav.familie.ba.sak.sikkerhet.SaksbehandlerContext
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.EnumSource
-import java.time.LocalDate
 import java.time.YearMonth
 
 class BrevServiceTest {
@@ -28,16 +23,15 @@ class BrevServiceTest {
     val endretUtbetalingAndelRepository = mockk<EndretUtbetalingAndelRepository>()
     val vedtaksbrevFellesfelterService = mockk<VedtaksbrevFellesfelterService>()
     val opprettGrunnlagOgSignaturDataService = mockk<OpprettGrunnlagOgSignaturDataService>()
+    val starttidspunktForUtbetalingstabellService = mockk<StarttidspunktForUtbetalingstabellService>()
 
     val brevService =
         BrevService(
-            persongrunnlagService = mockk(),
             simuleringService = mockk(),
             vedtaksperiodeService = vedtaksperiodeService,
             korrigertEtterbetalingService = mockk(),
             organisasjonService = mockk(),
             korrigertVedtakService = mockk(),
-            saksbehandlerContext = saksbehandlerContext,
             brevmalService = brevmalService,
             integrasjonClient = mockk(),
             andelTilkjentYtelseRepository = andelTilkjentYtelseRepository,
@@ -46,7 +40,8 @@ class BrevServiceTest {
             kompetanseRepository = mockk(),
             endretUtbetalingAndelRepository = endretUtbetalingAndelRepository,
             vedtaksbrevFellesfelterService = vedtaksbrevFellesfelterService,
-            opprettGrunnlagOgSignaturDataService = opprettGrunnlagOgSignaturDataService
+            opprettGrunnlagOgSignaturDataService = opprettGrunnlagOgSignaturDataService,
+            starttidspunktForUtbetalingstabellService = starttidspunktForUtbetalingstabellService,
         )
 
     @BeforeEach
@@ -119,87 +114,5 @@ class BrevServiceTest {
         val erLøpendeDifferanseUtbetalingPåBehandling = brevService.sjekkOmDetErLøpendeDifferanseUtbetalingPåBehandling(behandling)
 
         assertThat(erLøpendeDifferanseUtbetalingPåBehandling).isTrue()
-    }
-
-    @ParameterizedTest
-    @EnumSource(BehandlingÅrsak::class, mode = EnumSource.Mode.EXCLUDE, names = ["ÅRLIG_KONTROLL"])
-    fun `finnStarttidspunktForUtbetalingstabell returnerer endringstidspunkt for alle behandlingsårsaker utenom ÅRLIG_KONTROLL`(
-        behandlingÅrsak: BehandlingÅrsak,
-    ) {
-        every { vedtaksperiodeService.finnEndringstidspunktForBehandling(any()) } returns LocalDate.of(2020, 1, 1)
-
-        val behandling = lagBehandling(årsak = behandlingÅrsak)
-
-        val starttidspunkt = brevService.finnStarttidspunktForUtbetalingstabell(behandling)
-
-        assertThat(starttidspunkt).isEqualTo(LocalDate.of(2020, 1, 1))
-    }
-
-    @Test
-    fun `finnStarttidspunktForUtbetalingstabell returnerer endringstidspunkt for behandlingsårsak ÅRLIG_KONTROLL, dersom endringstidspunkt er tidligere enn 1 januar i fjor`() {
-        every { vedtaksperiodeService.finnEndringstidspunktForBehandling(any()) } returns LocalDate.of(2020, 1, 1)
-
-        val behandling = lagBehandling(årsak = BehandlingÅrsak.ÅRLIG_KONTROLL)
-
-        val starttidspunkt = brevService.finnStarttidspunktForUtbetalingstabell(behandling)
-
-        assertThat(starttidspunkt).isEqualTo(LocalDate.of(2020, 1, 1))
-    }
-
-    @Test
-    fun `finnStarttidspunktForUtbetalingstabell returnerer tidligst 1 januar i fjor for behandlingsårsak ÅRLIG_KONTROLL, dersom endringstidspunkt er TIDENES_ENDE`() {
-        every { vedtaksperiodeService.finnEndringstidspunktForBehandling(any()) } returns TIDENES_ENDE
-        every { endretUtbetalingAndelRepository.findByBehandlingId(any()) } returns emptyList()
-        every { andelTilkjentYtelseRepository.finnAndelerTilkjentYtelseForBehandling(any()) } returns
-            listOf(
-                lagAndelTilkjentYtelse(
-                    fom = YearMonth.of(2020, 1),
-                    tom = YearMonth.now().plusYears(1),
-                ),
-            )
-
-        val behandling = lagBehandling(årsak = BehandlingÅrsak.ÅRLIG_KONTROLL)
-
-        val starttidspunkt = brevService.finnStarttidspunktForUtbetalingstabell(behandling)
-
-        assertThat(starttidspunkt).isEqualTo(LocalDate.now().minusYears(1).withDayOfYear(1))
-    }
-
-    @Test
-    fun `finnStarttidspunktForUtbetalingstabell returnerer 1 januar i fjor for behandlingsårsak ÅRLIG_KONTROLL, selv om endringstidspunkt er senere`() {
-        every { vedtaksperiodeService.finnEndringstidspunktForBehandling(any()) } returns LocalDate.of(2024, 1, 1)
-        every { endretUtbetalingAndelRepository.findByBehandlingId(any()) } returns emptyList()
-        every { andelTilkjentYtelseRepository.finnAndelerTilkjentYtelseForBehandling(any()) } returns
-            listOf(
-                lagAndelTilkjentYtelse(
-                    fom = YearMonth.of(2020, 1),
-                    tom = YearMonth.now().plusYears(1),
-                ),
-            )
-
-        val behandling = lagBehandling(årsak = BehandlingÅrsak.ÅRLIG_KONTROLL)
-
-        val starttidspunkt = brevService.finnStarttidspunktForUtbetalingstabell(behandling)
-
-        assertThat(starttidspunkt).isEqualTo(LocalDate.now().minusYears(1).withDayOfYear(1))
-    }
-
-    @Test
-    fun `finnStarttidspunktForUtbetalingstabell returnerer første utbetalingstidspunkt ved ÅRLIG_KONTROLL dersom endringstidspunkt er TIDENES_ENDE og første utbetaling er etter 1 januar i fjor`() {
-        every { vedtaksperiodeService.finnEndringstidspunktForBehandling(any()) } returns TIDENES_ENDE
-        every { endretUtbetalingAndelRepository.findByBehandlingId(any()) } returns emptyList()
-        every { andelTilkjentYtelseRepository.finnAndelerTilkjentYtelseForBehandling(any()) } returns
-            listOf(
-                lagAndelTilkjentYtelse(
-                    fom = YearMonth.of(2024, 5),
-                    tom = YearMonth.now().plusYears(1),
-                ),
-            )
-
-        val behandling = lagBehandling(årsak = BehandlingÅrsak.ÅRLIG_KONTROLL)
-
-        val starttidspunkt = brevService.finnStarttidspunktForUtbetalingstabell(behandling)
-
-        assertThat(starttidspunkt).isEqualTo(LocalDate.of(2024, 5, 1))
     }
 }
